backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorama==0.4.6",
    "fastapi==0.104.1",
    "jinja2==3.1.2",
    "joblib==1.3.2",
    "mediapipe==0.10.7",
    "numpy==1.24.3",
    "openai==0.28",
    "opencv-python-headless==4.8.1.78",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "requests==2.31.0",
    "rich==13.7.0",
    "scikit-learn==1.3.2",
    "torch==2.1.0",
    "torchvision==0.16.0",
    "uvicorn[standard]==0.24.0",
    "websockets==12.0",
]
